Cleaned up code and ready for fresh start
