+++ GNSS/include/GNSS/gps_defs.hpp
@@ -2,7 +2,6 @@
 #define GPS_DEFS
 #include <cstdint>
 #include <cmath>
-#include "ephemeris.hpp"
 
 namespace gnss { namespace gps
 {
@@ -18,12 +17,6 @@ namespace gnss { namespace gps
     static double const pi = 3.1415926535898;
     static double const A_ref = 26559710; // meters
     static double const Omegad_ref = -2.6E-9; // semi-circles/second
-
-    static void parse_l1_ca(const uint32_t* words);
-    static void parse_subframe_3(const uint32_t* words);
-    static message_10 parse_L2_10(const uint32_t* words);
-    static message_11 parse_L2_11(const uint32_t* words);
-    static int32_t l1_get_split_data(const uint32_t* words, int position);
 } }
 
 #endif
